# models/book.go
package models


type Book struct {
	ID		 int    `json:"id" gorm:"primary_key"`
	Title	 string `json:"title" gorm:"not null"`
	AuthorID int 	`json:"Author_ID" gorm:"not null"`
	Author	 Author `json:"author" gorm:"foreignkey:AuthorID"`
}